refactor: grouping buttons component
