+++ src/components/buttons/index.ts
@@ -1,0 +1,4 @@
+import Button from "./Button";
+import ButtonIcon from "./ButtonIcon";
+
+export { Button, ButtonIcon };
