+++ src/screens/chat/sections/chat-detail/components/ChatDetailController.tsx
@@ -1,4 +1,4 @@
-import ButtonIcon from "@/components/ButtonIcon";
+import { ButtonIcon } from "@/components/buttons";
 import { BsThreeDots } from "react-icons/bs";
 
 const ChatDetailController = () => {
